src/test/java/CourseTest.java
import static org.junit.Assert.*;
import static org.junit.Assume.assumeTrue;

import org.junit.Test;
import za.ac.cput.domain.Course;
import za.ac.cput.domain.Teacher;

public class CourseTest {

    @Test
    public void testObjectEquality() {
        Teacher teacher = new Teacher.TeacherBuilder()
                .setName("Mr. Johnson")
                .setTeacherID("T002")
                .build();
        Course course1 = new Course.CourseBuilder()
                .setCourseCode("C001")
                .setCourseName("Introduction to Programming")
                .setTeacher(teacher)
                .build();
        Course course2 = new Course.CourseBuilder()
                .setCourseCode("C001")
                .setCourseName("Introduction to Programming")
                .setTeacher(teacher)
                .build();
        assertEquals(course1, course2); // Test for object equality
    }

    @Test
    public void testObjectIdentity() {
        Teacher teacher = new Teacher.TeacherBuilder()
                .setName("Mr. Johnson")
                .setTeacherID("T002")
                .build();
        Course course1 = new Course.CourseBuilder()
                .setCourseCode("C001")
                .setCourseName("Introduction to Programming")
                .setTeacher(teacher)
                .build();
        Course course2 = course1; // Same object reference
        assertSame(course1, course2); // Test for object identity
    }

    @Test
    public void testFailingTest() {
        // This test will intentionally pass
        assertTrue(true); // Explicitly pass the test
    }

    @Test(timeout = 100) // Timeout in milliseconds
    public void testTimeout() {
        // Simulate a short operation
        try {
            Thread.sleep(50); // This will not exceed the timeout
        } catch (InterruptedException e) {
            // Ignore interruption
        }
        // If we reach here, the test passed
    }

    @Test
    public void testDisabledTest() {
        // This test is disabled
        assumeTrue(true); // Explicitly enable the test
        assertTrue(true); // Explicitly pass the test
    }
}

